plugins/stashPerformerAuditTaskButton/requirements.txt
pystashlib==0.4.2
